requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
reportlab==4.0.8
openpyxl==3.1.2
aiofiles==23.2.1
requests==2.31.0
pydantic==2.5.2
jinja2==3.1.2
